giving deprecation a home
