+++ packages/server/modules/comments/domain.ts
@@ -1,0 +1,6 @@
+import type { Optional } from '@/modules/shared/helpers/typeHelper'
+import type { ExtendedComment } from "@/modules/comments/repositories/comments";
+
+export interface CommentsRepository {
+  getComment: (params: { id: string, userId?: string }) => Promise<Optional<ExtendedComment>>
+}
