+++ packages/server/modules/comments/services/retrieval/deprecated/index.ts
@@ -1,0 +1,1 @@
+export { getComments } from './getComments'
